+++ sdk/c4-client.ts
@@ -2,8 +2,8 @@
 // Generated from /openapi.json via src/openapi-sdk-gen.js.
 // Do not edit by hand — re-run `c4 openapi --sdk` to refresh.
 
-// Spec version: 1.10.61
-// Generated at: 2026-05-02T03:04:17.287Z
+// Spec version: 1.10.62
+// Generated at: 2026-05-02T03:16:49.940Z
 
 export interface postAuthLoginBody {
   user: string; /** Username */